refactor: split matrix thread write policy
